Don't use auth url to determine service protocol
Change the behavior of the command line glance client so that it no
longer assumes that the glance service always has the same protocol type
as the authentication service. Determine protocol from the given glance
service url. If no url is given, allow users to specify the protocol on
the command line with the --ssl option. If --ssl is not given, default
to using http.

Change-Id: I949c2ac429c67c3d1752a7b6fc840aa9e45fca8c